% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityplot.R
\name{.densityplot.adapor}
\alias{.densityplot.adapor}
\title{dispatch to different trellis object contructing function based on stack argument}
\usage{
.densityplot.adapor(x, data, stack = TRUE, plotType = "densityplot",
  ...)
}
\arguments{
\item{stack}{\code{logical} indicating whether to stack `name` on y axis}
}
\description{
dispatch to different trellis object contructing function based on stack argument
}
