% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyplot.R
\name{.process_flowFrame_overlay}
\alias{.process_flowFrame_overlay}
\title{convert a single flowFrame or a list of flowFrames to a list of flowSet}
\usage{
.process_flowFrame_overlay(overlay, sn)
}
\arguments{
\item{overlay}{flowFrame or a list of flowFrame objects}

\item{sn}{sample name}
}
\description{
convert a single flowFrame or a list of flowFrames to a list of flowSet
}
