% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyplot.R
\name{.process_overlay_flowSet}
\alias{.process_overlay_flowSet}
\title{extract the respective flowFrame from each flowSet based on the given sampleName}
\usage{
.process_overlay_flowSet(overlay, nm)
}
\arguments{
\item{overlay}{a list of flowSet}

\item{nm}{sample name}
}
\description{
extract the respective flowFrame from each flowSet based on the given sampleName
}
