% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyplot.R
\name{.xyplot.flowSet}
\alias{.xyplot.flowSet}
\title{\item{marker.only}{ \code{ligcal} specifies whether to show both channel and marker names }}
\usage{
.xyplot.flowSet(x, data, smooth = TRUE, filter = NULL,
  as.table = TRUE, prepanel = prepanel.xyplot.flowset,
  panel = panel.xyplot.flowset, xlab = NULL, ylab = NULL,
  par.settings = NULL, axis = axis.grid, defaultCond = "name",
  between = list(x = 0.2, y = 0.2), plotType = "xyplot",
  marker.only = FALSE, ...)
}
\description{
\item{marker.only}{ \code{ligcal} specifies whether to show both channel and marker names }
}
