% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flowViz-package.R
\docType{package}
\name{flowViz-package}
\alias{flowViz}
\alias{flowViz-package}
\title{Visualization for flow cytometry}
\description{
Functions and methods to visualize flow cytometry data. This package heavily
depends on the flowCore package.
}
\details{
\tabular{ll}{ Package: \tab flowViz\cr Type: \tab Package\cr Version: \tab
0.2.1\cr Date: \tab 2006-11-16\cr License: \tab Artistic\cr } Traditionally,
large parts of the analysis process of flow cytometry data has been mostly
qualitative. To this end, dedicated visualization techniques have been used
for both quality control and infernece of the data. This package provides a
number of different visualization tools for flow data.
}
\examples{
## examples go here
}
\author{
Maintainer: Florian Hahne <f.hahne@dkfz.de> Authors: T. Duong, B. Ellis, R.
Gentleman, F. Hahne, N. Le Meur, D. Sarkar, M. Tang
}
\seealso{
\code{\link[flowCore:flowCore-package]{flowCore}}
}
\keyword{package}

