\name{flowViz-package}
\alias{flowViz-package}
\alias{flowViz}
\docType{package}
\title{
  Visualization for flow cytometry
}
\description{
  Functions and methods to visualize flow cytometry data. This package
  heavily depends on the flowCore package.
}
\details{
\tabular{ll}{
Package: \tab flowViz\cr
Type: \tab Package\cr
Version: \tab 0.2.1\cr
Date: \tab 2006-11-16\cr
License: \tab Artistic\cr
}
Overview of packages goes here
}
\author{


  Maintainer: Florian Hahne <f.hahne@dkfz.de>
  Authors: T. Duong, B. Ellis, R. Gentleman, F. Hahne, N. Le Meur, D. Sarkar, M. Tang
}
\references{
references go here
}

\keyword{ package }
\seealso{
\code{\link[flowCore:flowCore-package]{flowCore}}
}
\examples{
## examples go here
}
